Fix integer formatting/casting

+++ packages/patrol/darwin/patrol/Sources/patrol/include/PatrolIntegrationTestIosRunner.h
@@ -44,8 +44,8 @@
     XCUIApplication *app = [[XCUIApplication alloc] init];                                                         \
     NSMutableDictionary<NSString *, NSString *> *environment =                                                     \
         [app.launchEnvironment mutableCopy] ?: [NSMutableDictionary dictionary];                                   \
-    environment[@"PATROL_TEST_SERVER_PORT"] = [NSString stringWithFormat:@"%d", server.boundTestPort];             \
-    environment[@"PATROL_APP_SERVER_PORT"] = [NSString stringWithFormat:@"%d", server.boundAppPort];               \
+    environment[@"PATROL_TEST_SERVER_PORT"] = [NSString stringWithFormat:@"%ld", (long)server.boundTestPort];      \
+    environment[@"PATROL_APP_SERVER_PORT"] = [NSString stringWithFormat:@"%ld", (long)server.boundAppPort];        \
     app.launchEnvironment = environment;                                                                            \
     [app launch];                                                                                                   \
   }                                                                                                                 \
@@ -241,7 +241,7 @@
     if (allDartTests == nil) {                                                                                      \
       [__test_class launchPatrolAppWithServer:server];                                                              \
       if (![__test_class waitForPatrolAppReadyWithServer:server]) {                                                \
-        XCTFail(@"Patrol app did not become ready on port %d", server.boundAppPort);                               \
+        XCTFail(@"Patrol app did not become ready on port %ld", (long)server.boundAppPort);                        \
         return @[];                                                                                                 \
       }                                                                                                             \
       __block NSArray<NSDictionary *> *listedTests = NULL;                                                          \
@@ -323,7 +323,7 @@
                                                                                                                     \
         [__test_class launchPatrolAppWithServer:server];                                                              \
         BOOL appReady = [__test_class waitForPatrolAppReadyWithServer:server];                                     \
-        XCTAssertTrue(appReady, @"Patrol app did not become ready on port %d", server.boundAppPort);              \
+        XCTAssertTrue(appReady, @"Patrol app did not become ready on port %ld", (long)server.boundAppPort);       \
         if (!appReady) {                                                                                            \
           return;                                                                                                   \
         }                                                                                                           \
